+++ code/algorithm-demo/src/com/noob/algorithm/base/template/MainForTree.java
@@ -1,0 +1,56 @@
+package com.noob.algorithm.base.template;
+
+
+import com.noob.algorithm.base.util.PrintListUtil;
+
+import java.util.ArrayList;
+import java.util.List;
+
+public class MainForTree {
+
+    // 初始化测试用例
+    public static List<int[]> initCase() {
+        List<int[]> caseList = new ArrayList<>();
+
+        // todo 补充测试用例集合
+        caseList.add(new int[]{1, 2, 3});
+
+        return caseList;
+    }
+
+    public static void testSolution1(List<int[]> caseList) {
+
+        for (int[] nums : caseList) {
+            // ① 调用算法
+
+
+            // ② 打印数据
+            PrintListUtil<String> printListUtil = new PrintListUtil<>();
+            // printListUtil.printMatrix(res);
+            System.out.println("********************");
+
+        }
+
+    }
+
+
+    public static void testSolution2(List<int[]> caseList) {
+
+    }
+
+
+    public static void main(String[] args) {
+
+        // 参数初始化
+        List<int[]> caseList = initCase();
+
+        System.out.println("---------- 算法①调用测试 ----------");
+
+        testSolution1(caseList);
+
+        System.out.println("---------- 算法②调用测试 ----------");
+
+        testSolution2(caseList);
+
+    }
+}
