+++ code/algorithm-demo/src/com/noob/algorithm/leetcode/q102/Solution1.java
@@ -1,7 +1,8 @@
 package com.noob.algorithm.leetcode.q102;
 
 
-import com.noob.algorithm.leetcode.structure.TreeNode;
+
+import com.noob.algorithm.base.dataStructure.tree.TreeNode;
 
 import java.util.ArrayList;
 import java.util.Deque;
