+++ code/algorithm-demo/src/com/noob/algorithm/leetcode/q437/Solution1.java
@@ -1,7 +1,7 @@
 package com.noob.algorithm.leetcode.q437;
 
 
-import com.noob.algorithm.leetcode.structure.TreeNode;
+import com.noob.algorithm.base.dataStructure.tree.TreeNode;
 
 /**
  * 437.路径总和III
